odroid-c1: add workaround for missing e2fsprogs-native tools
Dependency on e2fsprogs-native isn't automatically added in the
odroidc-sdimg, although the configuration seems to be correct. This
means that system mkfs.ext4 is used, but it lacks support for '-d'
option and results in the image creation error.

As a workaround, filesystem type is added to the configuration, which
adds e2fsprogs-native dependency.

Signed-off-by: Tomas Novotny <tomas.novotny@tbs-biometrics.com>